move validInputsFromPlan function]
